requirements.txt
fastapi
uvicorn
python-dotenv
pytest
pytest-asyncio
sqlalchemy
alembic
databases
pydantic
